requirements.txt
Flask-PyMongo==0.4.1
Flask-SocketIO==2.7.1
Flask==0.11.1
Jinja2==2.8
MarkupSafe==0.23
Werkzeug==0.11.11
click==6.6
eventlet==0.19.0
flake8==3.0.4
greenlet==0.4.10
itsdangerous==0.24
py==1.4.31
pylint==1.6.4
pymongo==3.3.0
pytest-cov==2.4.0
pytest==3.0.3
python-coveralls==2.9.0
python-dotenv==0.1.3
python-engineio==1.0.3
python-socketio==1.6.0
six==1.10.0
